refactor: renamed base package - updated kdoc for packages - renamed project
