+++ .github/workflows/build.yml
@@ -1,11 +1,11 @@
-name: build + sonar analysis
+name: Build
 
 on:
     push:
         branches:
             - main
         paths-ignore:
-            - 'docs/samples/**'
+            - 'docs/examples/**'
             - 'README.md'
     pull_request:
         types: [opened, synchronize, reopened]
@@ -15,26 +15,16 @@ jobs:
         runs-on: ubuntu-latest
 
         steps:
-            - uses: actions/checkout@v4
-              with:
-                  fetch-depth: 0
+            - name: Checkout code
+              uses: actions/checkout@v4
+
             - name: Set up JDK 21
               uses: actions/setup-java@v4
               with:
-                  java-version: 21
-                  distribution: 'temurin'
-            - name: Cache SonarQube packages
-              uses: actions/cache@v4
-              with:
-                  path: ~/.sonar/cache
-                  key: ${{ runner.os }}-sonar
-                  restore-keys: ${{ runner.os }}-sonar
-            - name: Cache Gradle packages
-              uses: actions/cache@v4
-              with:
-                  path: ~/.gradle/caches
-                  key: ${{ runner.os }}-gradle-${{ hashFiles('**/*.gradle') }}
-                  restore-keys: ${{ runner.os }}-gradle
+                java-version: 21
+                distribution: 'temurin'
+                cache: gradle
+
             - name: Build and analyze
               env:
                   SONAR_TOKEN: ${{ secrets.SONAR_TOKEN }}
